1.Number guessing game
This is a c++ code in which a random number(x) is generated every time and the user has to guess it. If the guessed number is smaller than x then the user is asked to guess the number again after giving him/her a hint that enter a greater number and same for a greater guessed number.